+++ src/components/Button/Button.tsx
@@ -7,12 +7,12 @@ import {
   ChakraProps,
   useMultiStyleConfig,
 } from "@chakra-ui/react";
+import { sizesArray } from "../../theme/sharedTypes";
 import React, { ButtonHTMLAttributes, forwardRef } from "react";
 
 import Icon from "../Icons/Icon";
 
 export const buttonElementTypeArray = ["submit", "button", "reset"] as const;
-export const buttonSizesArray = ["small", "medium", "large"] as const;
 export const buttonTypesArray = [
   "primary",
   "secondary",
@@ -23,7 +23,7 @@ export const buttonTypesArray = [
 ] as const;
 
 export type ButtonElementType = typeof buttonElementTypeArray[number];
-export type ButtonSizes = typeof buttonSizesArray[number];
+export type ButtonSizes = typeof sizesArray[number];
 export type ButtonTypes = typeof buttonTypesArray[number];
 
 export interface ButtonProps