+++ src/components/SubNav/SubNav.stories.tsx
@@ -5,12 +5,8 @@ import Hero from "../Hero/Hero";
 import Icon from "../Icons/Icon";
 import SkeletonLoader from "../SkeletonLoader/SkeletonLoader";
 
-import SubNav, {
-  SubNavButton,
-  SubNavLink,
-  actionBackgroundColorsArray,
-  highlightColorsArray,
-} from "./SubNav";
+import SubNav, { SubNavButton, SubNavLink } from "./SubNav";
+import { bgColorsArray, highlightColorsArray } from "../../theme/sharedTypes";
 
 const meta: Meta<typeof SubNav> = {
   title: "Components/Navigation/SubNav",
@@ -18,7 +14,7 @@ const meta: Meta<typeof SubNav> = {
   argTypes: {
     actionBackgroundColor: {
       control: "select",
-      options: actionBackgroundColorsArray,
+      options: bgColorsArray,
       defaultValue: { summary: "blogs" },
     },
     highlightColor: {