+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Adds the `useResponsiveSpacing` hook.
 - Adds docs for `Responsive Grid` Style Guide.
+- Consolidates message variants, sizes, highlight colors, and background colors arrays into shared constants
 
 ### Updates